move common code back into app
