--- platform/resources/assets/js/platform-bootstrap.js
@@ -1,9 +1,0 @@
-/**
- * Define the Vue filters.
- */
-require('filters');
-
-/**
- * Load the form utilities.
- */
-require('./forms/bootstrap');
