+++ resources/assets/js/filters/bootstrap.js
@@ -1,0 +1,9 @@
+/**
+ * Format the given date as a timestamp.
+ */
+require('./datetime');
+
+/**
+ * Format the given data as a human readable file size.
+ */
+require('./prettybytes');
