+++ resources/assets/js/filters/datetime.js
@@ -1,8 +1,5 @@
 window.moment = require('moment');
 
-/**
- * Format the given date as a timestamp.
- */
 Vue.filter('datetime', value => {
     return moment.utc(value).local().format('MMMM Do, YYYY h:mm A');
 });
